Doubles the elements of two arrays
In a class there are two privately declared integer arrays. Upon creation of an object both these arrays are initialized with different numbers of elements. Write a program that doubles the elements of both the arrays and assigns them to a single third array. Display the content of the third array. Take a different class for main() method.