Achieve inheritance with the help of interfaces
There are two interfaces, one has a method named sum and the other has a method named product. A class implements both these interfaces to display the sum and product of two integers taken from command line.
Java does not support multiple inheritance but the concept can be achieved with the help of interfaces. Illustrating this write a program for the above said problem.