Create System Generated Exception
handles two types of system generated exceptions: ArrayIndexOutOfBoundsException and ArithmeticException. When the string “AIOOBE” is passed from command line an object of the type ArrayIndexOutOfBoundsException is thrown, caught and printed. When the string “AE” is passed from command line an object of the typeArithmeticException is thrown, caught and printed. For any other string passed from command line, the output of the program displayed is “All is well”.